do consent mapping as i did for DC
