+++ crates/components-runtime/src/consent_mapping/mod.rs
@@ -1,7 +1,1 @@
 pub mod versions;
-
-wasmtime::component::bindgen!({
-    world: "consent-mapping",
-    path: "src/consent_mapping/wit",
-    async: true,
-});
