+++ crates/components-runtime/src/consent_mapping/wit/world.wit
@@ -1,5 +1,5 @@
 package edgee:native;
 
-world consent-mapping {
+world consent-mapping-v100 {
   export edgee:components/consent-mapping;
 }
